refactor the code, move generateToken to utils
