+++ controller/emergency.js
@@ -1,38 +1,13 @@
 const Patient = require('../models/patient');
 const Government = require('../models/government');
-const jwt = require('jsonwebtoken');
 const dotenv = require('dotenv');
 const chatsController = require('../controller/chats');
-const pushNotificationsController = require('./push_notifications');
+const {generateToken} = require('../general/utils');
 
 dotenv.config();
 
-const JWT_REMOTE_ACCESS_SECRET = process.env.JWT_REMOTE_ACCESS_SECRET;
-const JWT_REMOTE_ACCESS_EXPIRE = process.env.JWT_REMOTE_ACCESS_EXPIRE;
-
-// Function to generate a UUID
-function uuid() {
-    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
-        var r = Math.random() * 16 | 0,
-            v = c == 'x' ? r : (r & 0x3 | 0x8);
-        return v.toString(16);
-    });
-}
-
-function generateToken(patientUsername,doctorUsername){
-    const payload = {
-        patientUsername: patientUsername,
-        doctorUsername : doctorUsername,
-        uniqueId: uuid(), // Add a unique identifier to the payload
-    };
-
-    // Generate JWT token
-    const token = jwt.sign(payload, JWT_REMOTE_ACCESS_SECRET, { expiresIn:  JWT_REMOTE_ACCESS_EXPIRE});
-    return token;
-}
-
 const emergencyController = {
-    generateToken: async (req, res) => {
+    getEmergencyToken: async (req, res) => {
         const patient_ssn = req.query.patient_ssn;
         
         // check that the patient is in the database
