+++ controller/push_notifications.js
@@ -1,9 +1,8 @@
 const dotenv = require('dotenv');
-const jwt = require('jsonwebtoken');
 const Patient = require('../models/patient');
 const Firebase = require('../models/firebase');
-const File = require('../models/file');
 var admin = require("firebase-admin");
+const {generateToken} = require('../general/utils');
 
 dotenv.config();
 
@@ -14,30 +13,6 @@ admin.initializeApp({
   credential: admin.credential.cert(serviceAccount)
 });
 
-
-const JWT_REMOTE_ACCESS_SECRET = process.env.JWT_REMOTE_ACCESS_SECRET;
-const JWT_REMOTE_ACCESS_EXPIRE = process.env.JWT_REMOTE_ACCESS_EXPIRE;
-
-// Function to generate a UUID
-function uuid() {
-    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
-        var r = Math.random() * 16 | 0,
-            v = c == 'x' ? r : (r & 0x3 | 0x8);
-        return v.toString(16);
-    });
-}
-
-function generateToken(patientUsername,doctorUsername){
-    const payload = {
-        patientUsername: patientUsername,
-        doctorUsername : doctorUsername,
-        uniqueId: uuid(), // Add a unique identifier to the payload
-    };
-
-    // Generate JWT token
-    const token = jwt.sign(payload, JWT_REMOTE_ACCESS_SECRET, { expiresIn:  JWT_REMOTE_ACCESS_EXPIRE});
-    return token;
-}
 
 const pushNotificationsController = {
     giveAccess : async (req,res) => {
